requirements.txt
pyyaml==6.0
PyPDF2==2.8.1
pyttsx3
